Tabulator: moved Transformer, Recipe into own modules
