+++ aiida_jutools/io/tabulator/tabulator.py
@@ -12,230 +12,24 @@
 ###############################################################################
 """Tools for working with AiiDA IO: tabulation: Tabulator."""
 
+import abc as _abc
 import copy as _copy
 import json
 import json as _json
-import pandas as _pd
-import aiida.orm as _orm
-import numpy as _np
-from masci_tools.util import python_util as _masci_python_util
-import abc as _abc
 import typing as _typing
+
+import aiida.orm as _orm
+import pandas as _pd
+from masci_tools.util import python_util as _masci_python_util
+
 import aiida_jutools as _jutools
-
-
-class Transformer(_abc.ABC):
-    """Specify how to transformer an object's properties for use in :py:class:`Tabulator`.
-
-    To subclass, you have to implement the :py:meth:`~transformer` method.
-    """
-
-    @_abc.abstractmethod
-    def transform(self, keypath: _typing.Union[str, _typing.List[str]],
-                  value: _typing.Any,
-                  obj: _typing.Any = None,
-                  **kwargs) -> _typing.Tuple[_typing.Union[None, _typing.Any, dict], bool]:
-        """Specify how to transform properties, based on their keypath and type.
-
-        Extends :py:meth:`~.Transformer.transform`. See also its docstring.
-
-        This default transformer returns all property values unchanged, and so has no effect. To define
-        transformations, create a subclass and overwrite with a custom transform method.
-
-         Example: Say, a nested dictionary is passed. It has a property
-         `a_dict:{outputs:{last_calc_output_parameters:{attributes:{total_charge_per_atom:[...]`, which
-         is a numerical list. We would like the list, and its maximum tabulated as individual columns.
-         All other properties included in the tabulation shall be not transformed. We would write
-         that transformation rule like this.
-
-        .. code-block:: python
-
-           if keypath == ['outputs', 'last_calc_output_parameters', 'total_charge_per_atom']:
-               # assert isinstance(value, list) # optional
-               return {'total_charge_per_atom': value,
-                       'maximum_total_charge': max(value)}, True
-
-           return value, False
-
-        All kinds of transformation rules for all kinds of properties can be tailored in this way
-        to the specific use-case. Keep in mind that if a include list is used, the property (path) has
-        to be included in the include list.
-
-        If used in `aiida-jutools`: For accessing process node inputs and outputs Dict nodes properties:
-        first key in keypath is 'inputs' or 'outputs', the second is the input or output name that `node.outputs.`
-        displays on tab completion (which, under the hood, comes from the in- or outgoing link_label).
-
-        :param keypath: A list of keys, one per nesting level. First key in keypath is usually the object's attribute
-        name.
-        :param value: The value of the current property.
-        :param obj: Optionally, the object containing the property can be passed along. This enables to
-                    transform the current property value in combination with other property values.
-        :param kwargs: Additional keyword arguments for subclasses.
-        :return: A tuple (transformed_value:object, with_new_columns flag:bool). If the flag is False, this
-                 means the transformed output property has the same name as the input property (in/out referring
-                 here to the input/output of the transform method, not the input/output properties of the object). If
-                 the value is a dict, and the flag is set to True, this is understood such that new output
-                 properties were created from the input property, and the output value should be interpreted as
-                 {property_name: property_value}, possibly with one being the original property name. In a tabulator,
-                 new columns would be created for these new properties.
-        """
-        pass
-
-
-class DefaultTransformer(Transformer):
-    """Extends :py:class:`~Transformer`.
-
-    This default transformer does nothing (invariant operation): it just returns the value it received, along with
-    the new columns flag with value 'False'.
-    """
-
-    def transform(self,
-                  keypath: _typing.Union[str, _typing.List[str]],
-                  value: _typing.Any,
-                  obj: _typing.Any = None,
-                  **kwargs) -> _typing.Tuple[_typing.Union[None, _typing.Any, dict], bool]:
-        return value, False
-
-
-class Recipe(_abc.ABC):
-    def __init__(self,
-                 exclude_list: dict = None,
-                 include_list: dict = None,
-                 transformer: Transformer = None,
-                 **kwargs):
-        """Initialize a tabulator object.
-
-        The attributes :py:attr:`~.include_list` and :py:attr:`~.exclude_list` control whic properties
-        are to be tabulated. They may be set in a derived class definition, or at runtime on an instance.
-
-        Subclasses define the nature of the objects to be tabulated by making assumptions on their
-        property structure. That way, if both include and exclude list are empty, by default the 'complete'
-        set of properties of the objects will be tabulated, where the subclass defines the notion of 'complete'.
-
-        If neither exclude nor include list is given, the full set of properties according to implementation
-        will be tabulated.
-
-        Format of the include and exclude lists:
-
-
-
-        :param exclude_list: Optional list of properties to exclude. May be set later.
-        :param include_list: Optional list of properties to include. May be set later.
-        :param transform: Specifies special transformations for certain properties for tabulation.
-        :param kwargs: Additional keyword arguments for subclasses.
-        """
-        # note: for the in/ex lists, using the public setter here,
-        # to trigger conversion
-        self._exclude_list = exclude_list if exclude_list else {}
-        self._include_list = include_list if include_list else {}
-        self.transformer = transformer
-
-        self._scalar_types = (bool, int, float, str, complex)
-        self._nonscalar_types = (list, tuple, set, _np.ndarray)
-        self._nested_types = tuple([dict])
-        self._simple_types = tuple(
-            set(self._scalar_types).union(
-                set(self._nonscalar_types))
-        )
-
-    @property
-    def exclude_list(self) -> dict:
-        return self._exclude_list
-
-    @exclude_list.setter
-    def exclude_list(self, exclude_list: _typing.Union[dict, list]):
-        self._exclude_list = exclude_list
-        if isinstance(exclude_list, dict):
-            self._to_keypaths()
-
-    @property
-    def include_list(self) -> dict:
-        return self._include_list
-
-    @include_list.setter
-    def include_list(self, include_list: _typing.Union[dict, list]):
-        self._include_list = include_list
-        if isinstance(include_list, dict):
-            self._to_keypaths()
-
-    def _to_keypaths(self):
-        """Generate paths from a possibly nested dictionary.
-
-        This method can be used for handling include lists, exclude lists, and when writing
-        new :py:class:`~Transformer` transform methods.
-
-        List of paths to each value within the dict as tuples (path, value).
-
-        convert from with-List to with-None format for convert to keypaths
-
-        convert to keypaths (upper: done inside this one anyway)
-        """
-
-        def _to_keypaths_recursive(sub_dict: dict,
-                                   path: list = []):
-            paths = []
-            for k, v in sub_dict.items():
-                if isinstance(v, dict):
-                    paths += _to_keypaths_recursive(v, path + [k])
-                paths.append((path + [k], v))
-            return paths
-
-        for in_or_ex, a_dict in {'in': self._include_list,
-                                 'out': self._exclude_list}.items():
-
-            # precondition: not already keypaths format
-            is_list = isinstance(a_dict, list)
-            is_all_lists = is_list and all(isinstance(path, list) for path in a_dict)
-            if is_all_lists:
-                continue
-
-            # if empty, convert to empty list. if not empty, convert to keypaths
-            if not a_dict:
-                keypaths = []
-            else:
-                # convert from include list with-list format with-none format:
-                # same-level subkeys mentioned as list [k1,k2] -> dict {k1:None, k2:None}.
-                _a_dict = _masci_python_util.modify_dict(a_dict=a_dict,
-                                                         transform_value=lambda v: {k: None for k in v}
-                                                         if isinstance(v, list) else v,
-                                                         to_level=99)
-
-                keypaths = _to_keypaths_recursive(sub_dict=_a_dict,
-                                                  path=[])
-                # the result consists of sets of subpaths. For each subset, there is
-                # an additianal entry where the value contains the whole subdict from
-                # which the paths were generated. We are not interested in those duplicate
-                # entries, so remove them.
-                keypaths = [tup for tup in keypaths if not isinstance(tup[1], dict)]
-
-                # now list should be like [(path1, None), (path2, None), ...],
-                # or at least of type _typing.List[_typing.Tuple[list, _typing.Any]].
-                # check that. if not, something is wrong.
-                # otherwise, just return the paths.
-                if all(tup[1] is None for tup in keypaths):
-                    keypaths = [tup[0] for tup in keypaths]
-
-            # postcondition: keypaths format
-            is_list = isinstance(keypaths, list)
-            is_all_lists = is_list and all(isinstance(path, list) for path in keypaths)
-            if not is_all_lists:
-                raise TypeError(f"Could not generate keypaths of required type {_typing.List[list]} "
-                                f"from {in_or_ex}clude list. Either specified list in wrong format "
-                                f"(see class init docstring for examples), or list generated from "
-                                f"autolist stumbled over untreated special case for some unpacked "
-                                f"property.")
-
-            if in_or_ex == 'in':
-                self._include_list = keypaths
-            elif in_or_ex == 'out':
-                self._exclude_list = keypaths
 
 
 class Tabulator(_abc.ABC):
     """For tabulation of a collection of objects' (common) properties into a dict or dataframe."""
 
     def __init__(self,
-                 recipe: Recipe = None,
+                 recipe: _jutools.io.tabulator.Recipe = None,
                  **kwargs):
         """Initialize a tabulator object.
 
@@ -257,7 +51,7 @@ class Tabulator(_abc.ABC):
         :param kwargs: Additional keyword arguments for subclasses.
         """
         if not recipe:
-            recipe = Recipe()
+            recipe = _jutools.io.tabulator.Recipe()
         self.recipe = recipe
         self._table_types = []
         self._table = None
@@ -326,7 +120,7 @@ class NodeTabulator(Tabulator):
     """
 
     def __init__(self,
-                 recipe: Recipe = None,
+                 recipe: _jutools.io.tabulator.Recipe = None,
                  **kwargs):
         """Init node tabulator.
